pyproject.toml
[tool.poetry]
name = "llmcommitmsg"
version = "0.2.0"
description = "CLI tool to generate a commit message based on your Git diff - using an LLM"
authors = ["Michael Currin <18750745+MichaelCurrin@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.scripts]
lcm = "llmcommitmsg.__main__:main"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.58.1"
gitpython = "^3.1.43"


[tool.poetry.group.dev.dependencies]
ruff = "^0.8.4"
pytest-cov = "^6.2.1"
pytest = "^8.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
